Refactored code to combine the Core Registry and Registrar, into just a Registrar
