+++ contracts/interfaces/IRegistrar.sol
@@ -1,17 +1,22 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.7.3;
 
-import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
-
-interface IRegistrar is IERC721 {
+interface IRegistrar {
   // Emitted when a controller is removed
   event ControllerAdded(address indexed controller);
 
   // Emitted whenever a controller is removed
   event ControllerRemoved(address indexed controller);
 
-  // Emitted whenever a new name is registered
-  event NameRegistered(uint256 indexed id, address indexed owner);
+  // Emitted whenever a new domain is created
+  event DomainCreated(
+    uint256 indexed id,
+    string name,
+    uint256 indexed nameHash,
+    uint256 indexed parent
+  );
+
+  function domainExists(uint256 id) external view returns (bool);
 
   // Authorises a controller, who can register domains.
   function addController(address controller) external;
@@ -22,6 +27,9 @@ interface IRegistrar is IERC721 {
   // Whether or not a domain specific by an id is available.
   function available(uint256 id) external view returns (bool);
 
-  // Used by controller to register a new domain via id.
-  function register(uint256 id, address owner) external returns (uint256);
+  function registerDomain(
+    uint256 parent,
+    string memory name,
+    address domainOwner
+  ) external;
 }
